feat: upload file

+++ server/api/resource/RunProgramming.py
@@ -4,77 +4,112 @@ from flask_restful import Resource, Api
 from subprocess import PIPE
 import subprocess
 import time
+import os
+import shutil
+from werkzeug.utils import secure_filename
 
+ALLOWED_EXTENSIONS = {'sof', 'pgv'}
 app = Flask(__name__)
 api = Api(app)
 
 class RunProgramming(Resource):
     '''
-    上傳完PGV和SOF檔之後，傳兩個檔案的路徑和檔案名稱給這個api，之後就會燒程式進去DE0然後回傳law檔案回去
+    利用form-data傳資料
+    上傳完PGV和SOF檔之後，之後就會燒程式進去DE0然後回傳law檔案回去
     Inputs:
-        codePath:有PGV和SOF檔的檔案路徑
-        pgvName:向量檔的檔案名稱
-        sofName:sof檔的檔案名稱
+        Username:學生帳號
+        pgvFile:pgv檔案    <input type=file name=pgvFile>
+        sofFile:sof檔案    <input type=file name=sofFile>
     Outputs:
         law file:燒錄完的波型檔
     '''
     def get(self):
         return {'Msg': 'This is GET method!'}
     def post(self):
-        pgvDataPath = request.get_json()
-        print(pgvDataPath['codePath'])
-        batPath = "C:\\git-repos\\ours\\CloudLab\\server\\api\\common\\"
+        #設定資料夾路徑
+        LAFlag = 0
+        succFlag = 0
+        Username = request.form["Username"]
+        UPLOAD_FOLDER = "C:\\git-repos\\ours\\CloudLab\\server\\file\\" + Username
+        app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 
-        # batPath = "..\\common\\PG_run.bat"
-
-        ###write PG_run bat file
-        with open(pgvDataPath['codePath'] + "\\" + pgvDataPath['pgvName'],'r') as f:
-            data = f.read()
-
-            data = data.strip()
-            # print(data)
-            tmp = data.splitlines()
-
-            timeUnit = tmp[3]
-            timeUnit = timeUnit.strip()[5] #取得時間單位
-            # print(tmp)
-
-            #取得輸出波型需要幾秒
-            if(len(tmp[-1]) == 1):
-                wtime = tmp[-2]
-            else:
-                wtime = tmp[-1]
-            wtime = wtime[:wtime.rfind('>')].strip()
-
-            #算出需要多少ms輸出波型
-            if timeUnit == 'N' or timeUnit == 'n':
-                waveTime = float(wtime) * pow(10,-6)
-            if timeUnit == 'U' or timeUnit == 'u':
-                waveTime = float(wtime) * pow(10,-3)
-            if timeUnit == 'M' or timeUnit == 'm':
-                waveTime = float(wtime)
-
-            print("Time is {}ms".format(str(waveTime)))
-
-
-        with open(batPath + "PG_run.bat",'w') as fileWrite:
-            fileWrite.write("cd " + pgvDataPath['codePath'])
-            fileWrite.write("\nC:\\git-repos\\ours\\CloudLab\\server\\api\\common\\programming\\PG_run\\bin\\x86\\Debug\\PG_1.exe {} {}".format(pgvDataPath['pgvName'],str(waveTime)))
-        ###
+        #做檔案的副檔名檢查
+        def allowed_file(filename):
+            return '.' in filename and \
+            filename.rsplit('.', 1)[1] in ALLOWED_EXTENSIONS
         
-        ### write Quartus programming bat file
-        with open(batPath + "Programming_run.bat",'w') as fileWrite:
-            fileWrite.write("cd " + pgvDataPath['codePath'])
-            fileWrite.write("\nC:\\altera\\13.0\\quartus\\bin64\\quartus_pgm.exe -m JTAG -o p;{}".format(pgvDataPath['sofName']))
-        ###
+        def upload_file(fileKey):
+            file = request.files[fileKey]
+            if file and allowed_file(file.filename):        #檢查副檔名
+                filename = secure_filename(file.filename)    #避免Directory traversal attack
+                file.save(os.path.join(app.config['UPLOAD_FOLDER'], filename))   #存檔
+                return filename
 
-        ### write LA bat file
-        with open(batPath + "LA_run.bat",'w') as fileWrite:
-            fileWrite.write("cd " + pgvDataPath['codePath'])
-            fileWrite.write("\nC:\\git-repos\\ours\\CloudLab\\server\\api\\common\\programming\\LA_run\\bin\\Debug\\C_Sharp.exe {}".format("00857142"))
-        ###
-
+        # pgvDataPath = request.get_json()
         try:
+            #檢查資料夾存不存在，if not then create the folder
+            if not os.path.isdir(UPLOAD_FOLDER):
+                os.mkdir(UPLOAD_FOLDER)
+            else:
+                shutil.rmtree(UPLOAD_FOLDER)
+                os.mkdir(UPLOAD_FOLDER)
+
+            #上傳sof,pgv檔到files
+            sofName = upload_file("sofFile")
+            pgvName = upload_file("pgvFile")
+
+            # print(pgvDataPath['codePath'])
+            batPath = "C:\\git-repos\\ours\\CloudLab\\server\\api\\common\\"
+
+            # batPath = "..\\common\\PG_run.bat"
+
+            ###write PG_run bat file
+            with open(UPLOAD_FOLDER + "\\" + pgvName,'r') as f:
+                data = f.read()
+
+                data = data.strip()
+                # print(data)
+                tmp = data.splitlines()
+
+                timeUnit = tmp[3]
+                timeUnit = timeUnit.strip()[5] #取得時間單位
+                # print(tmp)
+
+                #取得輸出波型需要幾秒
+                if(len(tmp[-1]) == 1):
+                    wtime = tmp[-2]
+                else:
+                    wtime = tmp[-1]
+                wtime = wtime[:wtime.rfind('>')].strip()
+
+                #算出需要多少ms輸出波型
+                if timeUnit == 'N' or timeUnit == 'n':
+                    waveTime = float(wtime) * pow(10,-6)
+                if timeUnit == 'U' or timeUnit == 'u':
+                    waveTime = float(wtime) * pow(10,-3)
+                if timeUnit == 'M' or timeUnit == 'm':
+                    waveTime = float(wtime)
+
+                print("Time is {}ms".format(str(waveTime)))
+
+
+            with open(batPath + "PG_run.bat",'w') as fileWrite:
+                fileWrite.write("cd " + UPLOAD_FOLDER)
+                fileWrite.write("\nC:\\git-repos\\ours\\CloudLab\\server\\api\\common\\programming\\PG_run\\bin\\x86\\Debug\\PG_1.exe {} {}".format(pgvName,str(waveTime)))
+            ###
+            
+            ### write Quartus programming bat file
+            with open(batPath + "Programming_run.bat",'w') as fileWrite:
+                fileWrite.write("cd " + UPLOAD_FOLDER)
+                fileWrite.write("\nC:\\altera\\13.0\\quartus\\bin64\\quartus_pgm.exe -m JTAG -o p;{}".format(sofName))
+            ###
+
+            ### write LA bat file
+            with open(batPath + "LA_run.bat",'w') as fileWrite:
+                fileWrite.write("cd " + UPLOAD_FOLDER)
+                fileWrite.write("\nC:\\git-repos\\ours\\CloudLab\\server\\api\\common\\programming\\LA_run\\bin\\Debug\\C_Sharp.exe {}".format(Username))
+            ###
+
             ### programming the DE0 board
             print("programming process\n")
             programming_process = subprocess.Popen([batPath + "Programming_run.bat"],stdout = PIPE) # run Programming_run.bat
@@ -91,6 +126,7 @@ class RunProgramming(Resource):
 
             ###Run the logic analysis(LA)
             print("LA process\n")
+            LAFlag = 1
             LA_process = subprocess.Popen([batPath + "LA_run.bat"])
             time.sleep(30)
             ###
@@ -106,22 +142,12 @@ class RunProgramming(Resource):
             ###
 
             time.sleep(20) #sleep for LA
+            succFlag = 1
         except Exception as err:
             state = int(str(err)[5])
-            # print("state is : {}".format(state))
-            # if(state == 0):
-            #     programming_process.kill();
-            # else:
-            #     PG_process.kill();
-            # LA_process.kill();
             print(str(err))
-        if(programming_process.poll() is None):
-            print("kill the programming_process!")
-            programming_process.kill()
-        if(PG_process.poll() is None):
-            print("kill the PG_process!")
-            PG_process.kill()
-        if(LA_process.poll() is None):
+
+        if(LAFlag and LA_process.poll() is None):
             print("kill the LA_process!")
             LA_process.kill()
 
@@ -151,7 +177,10 @@ class RunProgramming(Resource):
         
 
         print("go to the program end!!\n")
-        return {'Msg': 'Running the program!'}
+        if(succFlag):
+            return {'Msg': 'Running the program!'}
+        else:
+            return {'Msg': 'Error occurred!'}
 
 api.add_resource(RunProgramming, '/api/RunProgramming')
